ci: :zap: Folder re-structured and types added for Preload and UI Window
