+++ src/electron/tsconfig.json
@@ -4,6 +4,7 @@
         "target": "esnext",
         "module": "nodenext",
         "outDir": "../../dist-electron",
-        "skipLibCheck": true
+        "skipLibCheck": true,
+        "types": ["../types/types"]
     }
 }
