+++ src/types/types.d.ts
@@ -1,0 +1,8 @@
+interface Window {
+    electronAPI: {
+        minimize: () => void;
+        maximize: () => void;
+        close: () => void;
+        selectFolder: () => Promise<string | null>;
+    };
+}
