+++ src/ui/components/image-box.tsx
