+++ src/ui/page/home.tsx
@@ -1,10 +1,9 @@
-import ButtonGroupRipple from '@/components/buttons/button-group';
-import SaveCard from '@/components/cards/save-card';
+import ButtonGroupRipple from '@/ui/components/buttons/button-group';
+import SaveCard from '@/ui/components/cards/save-card';
 
 const Index = () => {
     return (
         <div className="flex flex-col gap-4 pr-8">
-            
             <div className="bg-card mt-10 w-fit rounded p-2">
                 <ButtonGroupRipple />
             </div>
