+++ tsconfig.app.json
@@ -5,7 +5,7 @@
         "useDefineForClassFields": true,
         "lib": ["ES2022", "DOM", "DOM.Iterable"],
         "module": "ESNext",
-        "types": ["vite/client"],
+        "types": ["vite/client", "./src/types/types"],
         "skipLibCheck": true,
 
         /* Bundler mode */
